requirements/base.txt
Django==2.1.1
djangorestframework==3.8.2
django-mptt==0.9.1